little fix.

+++ src/main/java/com/example/leaverequest/service/LeaveRequestServiceImpl.java
@@ -14,6 +14,9 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
@@ -26,29 +29,19 @@ public class LeaveRequestServiceImpl implements LeaveRequestService
     LeaveRequestRepository leaveRequestRepository;
     
     
-    @Override
-    @Transactional
-    @PreAuthorize("isAuthenticated()")
-    public LeaveRequest createLeaveRequest(LeaveRequestDTO dto)
-    {
-        LeaveRequest leaveRequest = new LeaveRequest(dto);
-        leaveRequest.setStatus(Status.WAITINGAPPROVAL.getStatus());
-        if (leaveRequest.valid())
-        {
-            return leaveRequestRepository.save(leaveRequest);
-        }
-        else
-        {
-            throw new EntityBadInformationsException("Leave request informations are incorrect");
-        }
-    }
-    
     // Not used
     @Override
     @PreAuthorize("isAuthenticated()")
     public List<LeaveRequest> getAllLeaveRequests()
     {
         return leaveRequestRepository.findAll();
+    }
+
+    @Override
+    @PreAuthorize("isAuthenticated()")
+    public List<LeaveRequest> getAllLeaveRequestsNotRejected()
+    {
+        return leaveRequestRepository.findAllByStatusNotLike(Status.REJECTED.getStatus());
     }
     
     @Override
@@ -67,6 +60,29 @@ public class LeaveRequestServiceImpl implements LeaveRequestService
     
     @Override
     @PreAuthorize("isAuthenticated()")
+    public List<Date> getAllDisabledDatesByPersonId(long personId)
+    {
+        Calendar c = Calendar.getInstance();
+        c.setTime(new Date());
+        c.add(Calendar.MONTH, -2);
+        
+        List<LeaveRequest> leaveRequests = leaveRequestRepository.findAllByPersonIdAndLeaveFromIsAfter(personId, c.getTime());
+        List<Date> dates = new ArrayList<>();
+        for (LeaveRequest request : leaveRequests) {
+            Date start = request.getLeaveFrom();
+            Date end = request.getLeaveTo();
+            while (start.compareTo(end) <= 0) {
+                dates.add(start);
+                c.setTime(start);
+                c.add(Calendar.DAY_OF_MONTH, 1);
+                start = c.getTime();
+            }
+        }
+        return dates;
+    }
+    
+    @Override
+    @PreAuthorize("isAuthenticated()")
     public Page<LeaveRequest> getAllLeaveRequestsByStatus(String status, Pageable pageable)
     {
         return leaveRequestRepository.findAllByStatusLike(status, pageable);
@@ -77,6 +93,23 @@ public class LeaveRequestServiceImpl implements LeaveRequestService
     public LeaveRequest getLeaveRequestById(long id)
     {
         return leaveRequestRepository.findOne(id);
+    }
+    
+    @Override
+    @Transactional
+    @PreAuthorize("isAuthenticated()")
+    public LeaveRequest createLeaveRequest(LeaveRequestDTO dto)
+    {
+        LeaveRequest leaveRequest = new LeaveRequest(dto);
+        leaveRequest.setStatus(Status.WAITINGAPPROVAL.getStatus());
+        if (leaveRequest.valid(getAllDisabledDatesByPersonId(leaveRequest.getPersonId())))
+        {
+            return leaveRequestRepository.save(leaveRequest);
+        }
+        else
+        {
+            throw new EntityBadInformationsException("Leave request informations are incorrect");
+        }
     }
     
     @Override
@@ -139,7 +172,12 @@ public class LeaveRequestServiceImpl implements LeaveRequestService
     @PreAuthorize("isAuthenticated()")
     public void deleteLeaveRequest(long id)
     {
-        leaveRequestRepository.delete(id);
+        LeaveRequest leaveRequest = leaveRequestRepository.findOne(id);
+        if (leaveRequest.getStatus().equals(Status.WAITINGAPPROVAL.getStatus())) {
+            leaveRequestRepository.delete(id);
+        } else {
+            throw new EntityBadInformationsException("This leave request is already in approbation");
+        }
     }
     
     @Override
